renamed sum_functions to sum
